refactor: move tests inline with their modules
Move tests from the separate `rules/src/tests/` directory into idiomatic
`#[cfg(test)] mod tests` blocks within each source file:

- Board setup tests -> board.rs
- Standard rule pipeline tests -> standard.rs
- Food spawning tests -> food.rs
- Point/distance tests -> types.rs

Shared test helpers (make_snake, make_board) live in
types::test_helpers, accessible to all modules' test blocks.

All 43 tests pass, clippy and fmt clean.
