--- rules/src/tests/mod.rs
@@ -1,2 +1,0 @@
-mod board_tests;
-mod standard_tests;
